requirements.txt
streamlit
pandas
plotly_express
numpy
streamlit-extras
openpyxl
locale